Traveling Salesman solution and circling
Initial upload. travelingsalesman.py solves the Traveling Salesman problem in Python by using a Simulated Annealing Algorithm while keeping track of the best solution. This finds the shortest path between a randomly chosen selection of sights in a city. The points are plotted on a cartoon map (spbmap.png) and the final tour is overlaid over the image.

The file clusterfind.py chooses a circle that encloses a set of randomly generated points and plots it. Also plotted is the largest distance between two points. The algorithm for this problem is very weak since it calculates the distance between each combination of points (the complexity is therefore N!).